requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==4.2.7
django-bootstrap-v5==1.0.11
django-filter==23.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-taggit==5.0.1
factory-boy==3.3.0
Faker==25.0.1
flake8==7.0.0
fontawesomefree==6.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.1
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
